pyproject.toml
[project]
name = "promptlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.78.1",
    "python-dotenv>=1.1.0",
]
